[test] apply the new test style to common_clickhouse_srv
