+++ common/clickhouse-srv/src/types/column/column_data.rs
@@ -23,9 +23,9 @@ use crate::types::SqlType;
 use crate::types::Value;
 use crate::types::ValueRef;
 
-pub(crate) type ArcColumnData = Arc<dyn ColumnData + Send + Sync>;
+pub type ArcColumnData = Arc<dyn ColumnData + Send + Sync>;
 
-pub(crate) type BoxColumnData = Box<dyn ColumnData + Send + Sync>;
+pub type BoxColumnData = Box<dyn ColumnData + Send + Sync>;
 
 pub trait ColumnData {
     fn sql_type(&self) -> SqlType;
@@ -45,7 +45,7 @@ pub trait ColumnData {
     }
 }
 
-pub(crate) trait ColumnDataExt {
+pub trait ColumnDataExt {
     fn append<T: convert::Into<Value>>(&mut self, value: T);
 }
