+++ common/clickhouse-srv/src/types/column/factory.rs
@@ -203,7 +203,7 @@ impl dyn ColumnData {
     }
 }
 
-fn parse_fixed_string(source: &str) -> Option<usize> {
+pub fn parse_fixed_string(source: &str) -> Option<usize> {
     if !source.starts_with("FixedString") {
         return None;
     }
@@ -215,7 +215,7 @@ fn parse_fixed_string(source: &str) -> Option<usize> {
     }
 }
 
-fn parse_nullable_type(source: &str) -> Option<&str> {
+pub fn parse_nullable_type(source: &str) -> Option<&str> {
     if !source.starts_with("Nullable") {
         return None;
     }
@@ -229,7 +229,7 @@ fn parse_nullable_type(source: &str) -> Option<&str> {
     Some(inner_type)
 }
 
-fn parse_array_type(source: &str) -> Option<&str> {
+pub fn parse_array_type(source: &str) -> Option<&str> {
     if !source.starts_with("Array") {
         return None;
     }
@@ -238,7 +238,7 @@ fn parse_array_type(source: &str) -> Option<&str> {
     Some(inner_type)
 }
 
-fn parse_decimal(source: &str) -> Option<(u8, u8, NoBits)> {
+pub fn parse_decimal(source: &str) -> Option<(u8, u8, NoBits)> {
     if source.len() < 12 {
         return None;
     }
@@ -317,12 +317,12 @@ fn parse_decimal(source: &str) -> Option<(u8, u8, NoBits)> {
     }
 }
 
-enum EnumSize {
+pub enum EnumSize {
     Enum8,
     Enum16,
 }
 
-fn parse_enum8(input: &str) -> Option<Vec<(String, i8)>> {
+pub fn parse_enum8(input: &str) -> Option<Vec<(String, i8)>> {
     match parse_enum(EnumSize::Enum8, input) {
         Some(result) => {
             let res: Vec<(String, i8)> = result
@@ -334,11 +334,11 @@ fn parse_enum8(input: &str) -> Option<Vec<(String, i8)>> {
         None => None,
     }
 }
-fn parse_enum16(input: &str) -> Option<Vec<(String, i16)>> {
+pub fn parse_enum16(input: &str) -> Option<Vec<(String, i16)>> {
     parse_enum(EnumSize::Enum16, input)
 }
 
-fn parse_enum(size: EnumSize, input: &str) -> Option<Vec<(String, i16)>> {
+pub fn parse_enum(size: EnumSize, input: &str) -> Option<Vec<(String, i16)>> {
     let size = match size {
         EnumSize::Enum8 => "Enum8",
         EnumSize::Enum16 => "Enum16",
@@ -385,7 +385,7 @@ fn parse_enum(size: EnumSize, input: &str) -> Option<Vec<(String, i16)>> {
     }
 }
 
-fn parse_date_time64(source: &str) -> Option<(u32, Option<String>)> {
+pub fn parse_date_time64(source: &str) -> Option<(u32, Option<String>)> {
     let integer = many1::<String, _, _>(digit()).and_then(|digits| {
         digits
             .parse::<u32>()
@@ -420,201 +420,9 @@ fn parse_date_time64(source: &str) -> Option<(u32, Option<String>)> {
     }
 }
 
-fn get_timezone(timezone: &Option<String>, tz: Tz) -> Result<Tz> {
+pub fn get_timezone(timezone: &Option<String>, tz: Tz) -> Result<Tz> {
     match timezone {
         None => Ok(tz),
         Some(t) => Ok(t.parse()?),
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn test_parse_decimal() {
-        assert_eq!(parse_decimal("Decimal(9, 4)"), Some((9, 4, NoBits::N32)));
-        assert_eq!(parse_decimal("Decimal(10, 4)"), Some((10, 4, NoBits::N64)));
-        assert_eq!(parse_decimal("Decimal(20, 4)"), None);
-        assert_eq!(parse_decimal("Decimal(2000, 4)"), None);
-        assert_eq!(parse_decimal("Decimal(3, 4)"), None);
-        assert_eq!(parse_decimal("Decimal(20, -4)"), None);
-        assert_eq!(parse_decimal("Decimal(0)"), None);
-        assert_eq!(parse_decimal("Decimal(1, 2, 3)"), None);
-        assert_eq!(parse_decimal("Decimal64(9)"), Some((18, 9, NoBits::N64)));
-    }
-
-    #[test]
-    fn test_parse_array_type() {
-        assert_eq!(parse_array_type("Array(UInt8)"), Some("UInt8"));
-    }
-
-    #[test]
-    fn test_parse_nullable_type() {
-        assert_eq!(parse_nullable_type("Nullable(Int8)"), Some("Int8"));
-        assert_eq!(parse_nullable_type("Int8"), None);
-        assert_eq!(parse_nullable_type("Nullable(Nullable(Int8))"), None);
-    }
-
-    #[test]
-    fn test_parse_fixed_string() {
-        assert_eq!(parse_fixed_string("FixedString(8)"), Some(8_usize));
-        assert_eq!(parse_fixed_string("FixedString(zz)"), None);
-        assert_eq!(parse_fixed_string("Int8"), None);
-    }
-
-    #[test]
-    fn test_parse_enum8() {
-        let enum8 = "Enum8 ('a' = 1, 'b' = 2)";
-
-        let res = parse_enum8(enum8).unwrap();
-        assert_eq!(res, vec![("a".to_owned(), 1), ("b".to_owned(), 2)])
-    }
-    #[test]
-    fn test_parse_enum16_special_chars() {
-        let enum16 = "Enum16('a_' = -128, 'b&' = 0)";
-
-        let res = parse_enum16(enum16).unwrap();
-        assert_eq!(res, vec![("a_".to_owned(), -128), ("b&".to_owned(), 0)])
-    }
-
-    #[test]
-    fn test_parse_enum8_single() {
-        let enum8 = "Enum8 ('a' = 1)";
-
-        let res = parse_enum8(enum8).unwrap();
-        assert_eq!(res, vec![("a".to_owned(), 1)])
-    }
-
-    #[test]
-    fn test_parse_enum8_empty_id() {
-        let enum8 = "Enum8 ('' = 1, '' = 2)";
-
-        let res = parse_enum8(enum8).unwrap();
-        assert_eq!(res, vec![("".to_owned(), 1), ("".to_owned(), 2)])
-    }
-
-    #[test]
-    fn test_parse_enum8_single_empty_id() {
-        let enum8 = "Enum8 ('' = 1)";
-
-        let res = parse_enum8(enum8).unwrap();
-        assert_eq!(res, vec![("".to_owned(), 1)])
-    }
-
-    #[test]
-    fn test_parse_enum8_extra_comma() {
-        let enum8 = "Enum8 ('a' = 1, 'b' = 2,)";
-
-        assert!(dbg!(parse_enum8(enum8)).is_none());
-    }
-
-    #[test]
-    fn test_parse_enum8_empty() {
-        let enum8 = "Enum8 ()";
-
-        assert!(dbg!(parse_enum8(enum8)).is_none());
-    }
-
-    #[test]
-    fn test_parse_enum8_no_value() {
-        let enum8 = "Enum8 ('a' =)";
-
-        assert!(dbg!(parse_enum8(enum8)).is_none());
-    }
-
-    #[test]
-    fn test_parse_enum8_no_ident() {
-        let enum8 = "Enum8 ( = 1)";
-
-        assert!(dbg!(parse_enum8(enum8)).is_none());
-    }
-
-    #[test]
-    fn test_parse_enum8_starting_comma() {
-        let enum8 = "Enum8 ( , 'a' = 1)";
-
-        assert!(dbg!(parse_enum8(enum8)).is_none());
-    }
-
-    #[test]
-    fn test_parse_enum16() {
-        let enum16 = "Enum16 ('a' = 1, 'b' = 2)";
-
-        let res = parse_enum16(enum16).unwrap();
-        assert_eq!(res, vec![("a".to_owned(), 1), ("b".to_owned(), 2)])
-    }
-
-    #[test]
-    fn test_parse_enum16_single() {
-        let enum16 = "Enum16 ('a' = 1)";
-
-        let res = parse_enum16(enum16).unwrap();
-        assert_eq!(res, vec![("a".to_owned(), 1)])
-    }
-
-    #[test]
-    fn test_parse_enum16_empty_id() {
-        let enum16 = "Enum16 ('' = 1, '' = 2)";
-
-        let res = parse_enum16(enum16).unwrap();
-        assert_eq!(res, vec![("".to_owned(), 1), ("".to_owned(), 2)])
-    }
-
-    #[test]
-    fn test_parse_enum16_single_empty_id() {
-        let enum16 = "Enum16 ('' = 1)";
-
-        let res = parse_enum16(enum16).unwrap();
-        assert_eq!(res, vec![("".to_owned(), 1)])
-    }
-
-    #[test]
-    fn test_parse_enum16_extra_comma() {
-        let enum16 = "Enum16 ('a' = 1, 'b' = 2,)";
-
-        assert!(dbg!(parse_enum16(enum16)).is_none());
-    }
-
-    #[test]
-    fn test_parse_enum16_empty() {
-        let enum16 = "Enum16 ()";
-
-        assert!(dbg!(parse_enum16(enum16)).is_none());
-    }
-
-    #[test]
-    fn test_parse_enum16_no_value() {
-        let enum16 = "Enum16 ('a' =)";
-
-        assert!(dbg!(parse_enum16(enum16)).is_none());
-    }
-
-    #[test]
-    fn test_parse_enum16_no_ident() {
-        let enum16 = "Enum16 ( = 1)";
-
-        assert!(dbg!(parse_enum16(enum16)).is_none());
-    }
-
-    #[test]
-    fn test_parse_enum16_starting_comma() {
-        let enum16 = "Enum16 ( , 'a' = 1)";
-
-        assert!(dbg!(parse_enum16(enum16)).is_none());
-    }
-
-    #[test]
-    fn test_parse_date_time64() {
-        let source = " DateTime64 ( 3 , 'Europe/Moscow' )";
-        let res = parse_date_time64(source).unwrap();
-        assert_eq!(res, (3, Some("Europe/Moscow".to_string())))
-    }
-
-    #[test]
-    fn test_parse_date_time64_without_timezone() {
-        let source = " DateTime64( 5 )";
-        let res = parse_date_time64(source).unwrap();
-        assert_eq!(res, (5, None))
     }
 }
