+++ common/clickhouse-srv/src/types/column/iter.rs
@@ -999,25 +999,3 @@ where T: Iterable<'a, Simple> + 'a
         })
     }
 }
-
-#[cfg(test)]
-mod test {
-    use crate::types::Block;
-
-    #[test]
-    fn test_complex_iter() {
-        let lo = Block::new().column("?", vec![1_u32, 2]);
-        let hi = Block::new().column("?", vec![3_u32, 4, 5]);
-
-        let block = Block::concat(&[lo, hi]);
-
-        let columns = block.columns()[0].iter::<u32>().unwrap();
-        let actual: Vec<_> = columns.collect();
-        assert_eq!(actual, vec![&1_u32, &2, &3, &4, &5]);
-
-        let null_u32 = Block::new().column("?", vec![Some(3_u32), Some(4), None]);
-        let columns = null_u32.columns()[0].iter::<Option<u32>>().unwrap();
-        let actual: Vec<_> = columns.collect();
-        assert_eq!(actual, vec![Some(&3_u32), Some(&4), None]);
-    }
-}
