+++ common/clickhouse-srv/tests/it/types/column/array.rs
@@ -12,10 +12,23 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod block;
-mod column;
-mod decimal;
-mod from_sql;
-mod opions;
-mod value;
-mod value_ref;
+use std::io::Cursor;
+
+use chrono_tz::Tz;
+use common_clickhouse_srv::binary::Encoder;
+use common_clickhouse_srv::types::*;
+
+#[test]
+fn test_write_and_read() {
+    let block = Block::<Simple>::new().column("vals", vec![vec![7_u32, 8], vec![9, 1, 2], vec![
+        3, 4, 5, 6,
+    ]]);
+
+    let mut encoder = Encoder::new();
+    block.write(&mut encoder, false);
+
+    let mut reader = Cursor::new(encoder.get_buffer_ref());
+    let rblock = Block::load(&mut reader, Tz::Zulu, false).unwrap();
+
+    assert_eq!(block, rblock);
+}
